--- a/bloomfilter_test.go
+++ b/bloomfilter_test.go
@@ -125,6 +125,9 @@ func TestAddX10kX5(t *testing.T) {
 		b1.Add(v)
 		b2.AddHash(v.Sum64())
 		verify()
+		if !b2.Contains(v) {
+			t.Fatal("contain error")
+		}
 	}
 }
 func BenchmarkContains1kX10kX5(b *testing.B) {
@@ -140,6 +143,7 @@ func BenchmarkContains1kX10kX5(b *testing.B) {
 }
 
 func BenchmarkContains100kX10BX20(b *testing.B) {
+	rand.Seed(1337)
 	b.StopTimer()
 	bf, _ := New(10*1000*1000*1000, 20)
 	for i := 0; i < 100*1000; i++ {