```
name                   old time/op    new time/op    delta
AddX10kX5/add-10kx5-6     152ns ± 2%     116ns ± 1%  -23.83%  (p=0.000 n=9+9)

name                   old alloc/op   new alloc/op   delta
AddX10kX5/add-10kx5-6     56.0B ± 0%      8.0B ± 0%  -85.71%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
AddX10kX5/add-10kx5-6      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=10+10)
```
```
name                   old time/op    new time/op    delta
Contains1kX10kX5-6        167ns ±15%     113ns ± 2%  -32.16%  (p=0.000 n=9+9)
Contains100kX10BX20-6    1.15µs ± 6%    0.79µs ±11%  -31.64%  (p=0.000 n=10+9)

name                   old alloc/op   new alloc/op   delta
Contains1kX10kX5-6        56.0B ± 0%      8.0B ± 0%  -85.71%  (p=0.000 n=10+10)
Contains100kX10BX20-6      168B ± 0%        8B ± 0%  -95.24%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
Contains1kX10kX5-6         2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=10+10)
Contains100kX10BX20-6      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=10+10)
```